refactor : application.properites 수정
