+++ src/main/resources/application.properties
@@ -1,1 +1,7 @@
 spring.application.name=spring-gift
+spring.h2.console.enabled=true
+spring.datasource.url=jdbc:h2:mem:test
+spring.datasource.name=sa
+spring.sql.init.schema-locations=classpath:/script/schema.sql
+spring.sql.init.data-locations=classpath:/script/data.sql
+spring.sql.init.mode=always
